Revert "move userMsgResult2Str() to conversion.cpp" to fix build failed
This reverts commit 9bdbcc8295bdffaddcc6aa2be0cd42666c53a6c4.
